Move LazyFrame to subfolder
