+++ py-polars/polars/internals/__init__.py
@@ -23,7 +23,6 @@ from polars.internals.io import (
     read_ipc_schema,
     read_parquet_schema,
 )
-from polars.internals.lazy_frame import LazyFrame, wrap_ldf
 from polars.internals.lazy_functions import (
     all,
     arg_where,
@@ -35,6 +34,7 @@ from polars.internals.lazy_functions import (
     lit,
     select,
 )
+from polars.internals.lazyframe import LazyFrame, wrap_ldf
 from polars.internals.series import Series, wrap_s
 from polars.internals.whenthen import when  # used in expr.clip()
