+++ py-polars/polars/internals/lazyframe/__init__.py
@@ -1,0 +1,6 @@
+from polars.internals.lazyframe.frame import LazyFrame, wrap_ldf
+
+__all__ = [
+    "LazyFrame",
+    "wrap_ldf",
+]
